fix data corruption on undoing some moves
When undoing a capture for the -1 player (AI) GameState was incorrectly
recovering the previous state of the game.

Also added a state generation function which takes human readable input
and creates a GameState object based on it. For now it comes in handy in
testing, but later might be useful for example for generating the
initial state of the game in a clean way.