fix inconsistent player representation
In some places -1 was used to describe the AI player and 1 for the human
player, whereas in others it was in reverse. From now on:
1 is human player
-1 is AI player

Regarding coordinates of the board:
Human player starts at low Y values and the AI starts at hight Y values
on the board. This is necessary to establish, because pieces might be
only able to move "forwards" during the game.