fix captures through walls
A bug caused the moves generator (successor function) to generate
captures of pieces that were obscured by other pieces. Therefore if two
pieces stood in front of one capable of capturing in range of 1..2
forwards, moves for capture of each were generated instead of only
generating the move for capture of the first piece.