add interaction with end user
Introduced EngineQuery and EngineResponse classes for convenient use of
the engine, also exposed the engine via a rest controller, so it can now
be used by front end apps.

Also temporarily made board public in GameState. This solves an issue
with Jackson not including the board in the JSON representation.