code cleanup
Moved files between packages in order to make the structure more
logical. Successor function now resides in a new "search" package, which
also contains other typical search-algorithm parts, like negamax and
evaluation.

Files most crucial for interaction of the engine with code using it are
kept in the "engine" package directly.

Replaced JVM'a Math.max and Math.abs with kotlin stdlib's max and abs.
This will be helpful should the project be turned into a kotlin's
multiplatform library in the future.