use integer sorting for move ordering
Replaced usual n log n sorting with a specialized, integer sorting
function tailored for this use case. This allowed for approximately
doubling of the performance of the entire engine except for cases where
there is a lot of self captures possible. For example  in the beginning
of a new game. This can be addressed by no longer sorting self captures
below moves with no capture. Something to consider.

Also replaced Lists with MutableLists, in order to prevent unnecessary
object creation. Performance impact is small.